+++ src/application.ts
@@ -2,7 +2,7 @@
 
 import { parseArguments, parseOptions } from "./parsing.ts";
 import { deepFreeze, maxLength, padStringTo } from "./utilities.ts";
-import { Application, ApplicationSpec, Command, Option } from "./types.ts";
+import { Action, Application, ApplicationSpec, Command, Option } from "./types.ts";
 import regexes from "./regexes.ts";
 
 // ============================== Application ============================== //
@@ -18,6 +18,9 @@ export function application(spec?: ApplicationSpec): Application {
 
 	const _commands: Command[] = [];
 	const _options: Option[] = [];
+
+	let _arguments: string;
+	let _action: Action;
 
 	// -------------------- Option & Command Validation -------------------- //
 
@@ -144,6 +147,14 @@ export function application(spec?: ApplicationSpec): Application {
 
 	// --------------------------- Public Methods --------------------------- //
 
+	const action = (act: Action): void => {
+		_action = act;
+	};
+
+	const args = (argsDefinition: string): void => {
+		_arguments = argsDefinition;
+	};
+
 	const command = (command: Command): void => {
 		_validateCommand(command);
 		_commands.push(command);
@@ -169,9 +180,21 @@ export function application(spec?: ApplicationSpec): Application {
 			return;
 		}
 
-		// This is seperate from option flag because intellisense
+		// A hack to remove all options from input, due to a
+		// limitation with iterating over the array instead
+		const providedArgs = (command ? input.slice(1) : input).join(" ").replace(
+			regexes.optionParse,
+			"",
+		).split(" ");
+
 		if (!command) {
-			if (_helpOptionEnabled) _help();
+			if (!_action && _helpOptionEnabled) {
+				_help();
+				return;
+			}
+
+			// Run command action if no command has ran
+			_action({ arguments: parseArguments(_arguments || "", providedArgs), options: options });
 			return;
 		}
 
@@ -182,9 +205,7 @@ export function application(spec?: ApplicationSpec): Application {
 
 		const args = parseArguments(
 			command.arguments || "",
-			// A hack to remove all options from input, due to a
-			// limitation with iterating over the array instead
-			(command ? input.slice(1) : input).join(" ").replace(regexes.optionParse, "").split(" "),
+			providedArgs,
 		);
 
 		command.action({ arguments: args, options: options });
@@ -193,6 +214,8 @@ export function application(spec?: ApplicationSpec): Application {
 	// ------------------------ Return frozen object ------------------------ //
 
 	return deepFreeze({
+		action,
+		args,
 		command,
 		commands,
 		options,